% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateTwoCauseFineGrayModel.R
\name{find_mu_c}
\alias{find_mu_c}
\title{Find the mean of censoring times}
\usage{
find_mu_c(c_mu, censor, cens.times, event.times, f.event, n)
}
\arguments{
\item{c_mu}{A numeric value representing the shift to be applied to censoring times.}

\item{censor}{A numeric value representing the desired censoring proportion.}

\item{cens.times}{A vector of censoring times.}

\item{event.times}{A vector of event times.}

\item{f.event}{A vector representing the event type.}

\item{n}{The number of observations.}
}
\value{
A numeric value representing the difference between the observed censoring
proportion and the desired censoring proportion. This is used for root-finding.
}
\description{
This function is an internal helper used to find the shift `c_mu` needed
to achieve a desired censoring proportion.
}
