% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.concordance_metric.R
\name{pam.concordance_metric}
\alias{pam.concordance_metric}
\title{Compute the Concordance Statistic for Survival Data}
\usage{
pam.concordance_metric(
  predicted_time,
  survival_time,
  status,
  weight = "H",
  input_tau = NULL
)
}
\arguments{
\item{predicted_time}{Numeric vector of predicted survival times.}

\item{survival_time}{Numeric vector of observed survival times.}

\item{status}{Numeric vector indicating censoring status (1 = event occurred, 0 = censored).}

\item{weight}{Character string specifying the weighting method to use. Options are:
- `"H"`: Harrell's C-index.
- `"U"`: Unweighted C-index with KM survival probability.
- `"U_tau"`: Unweighted C-index truncated at \code{input_tau}.}

\item{input_tau}{Optional numeric value specifying the truncation point for \code{"U_tau"} weighting.}
}
\value{
A numeric value representing the computed C-index. 
A value closer to 1 indicates better model performance, while a value closer to 0.5 
indicates no discriminative power.
}
\description{
This function calculates the Concordance Index (C-index), a performance metric 
used to evaluate survival prediction models. The C-index measures the ability 
of a model to correctly rank predicted survival times with observed survival outcomes.
}
\details{
The function performs pairwise comparisons of predicted survival times and observed outcomes. 
For weighted calculations (`"U"` and `"U_tau"`), the Kaplan-Meier survival probability 
is used to weight the pairs. The `"U_tau"` method additionally truncates comparisons 
at a specified time point (\code{input_tau}).
}
\examples{
library(PAmeasure)
predicted_time <- c(2.5, 3.2, 1.8, 4.1)
survival_time <- c(3, 4, 2, 5)
status <- c(1, 0, 1, 1)
pam.concordance_metric(predicted_time, survival_time, status, weight = "H")

}
\references{
F Harrell, R Califf, D Pryor, K Lee and R Rosati, Evaluating the yield of medical tests, J Am Medical Assoc, 1982.

R Peto and J Peto, Asymptotically efficient rank invariant test procedures (with discussion), J Royal Stat Soc A, 1972.

M Schemper, Cox analysis of survival data with non-proportional hazard functions, The Statistician, 1992.

H Uno, T Cai, M Pencina, R D'Agnostino and Lj Wei, On the C-statistics for evaluating overall adequacy of risk prediction procedures with censored survival data, Statistics in Medicine, 2011.

Therneau, T. M., Lumley, T., Atkinson, E., Crowson, C. (2024). survival: Survival Analysis. 
R package version 3.7-0. DOI: \doi{10.32614/CRAN.package.survival}. Available at \url{https://CRAN.R-project.org/package=survival}.
}
