% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.metrics_summary_predicted.R
\name{pam.metrics_summary_predicted}
\alias{pam.metrics_summary_predicted}
\title{Summary of Performance Metrics for Predicted Survival Data}
\usage{
pam.metrics_summary_predicted(
  predicted_data,
  survival_time,
  metric,
  status = NULL,
  tau = NULL,
  t_star = NULL,
  start_time = NULL
)
}
\arguments{
\item{predicted_data}{A numeric vector of predicted survival probabilities or scores.}

\item{survival_time}{A numeric vector of observed survival times.}

\item{status}{A numeric or logical vector indicating event status (1 for event, 0 for censoring).}

\item{tau}{An optional numeric value for restricted time horizon. Default is NULL.}

\item{t_star}{An optional numeric value specifying the time point for certain metrics. Default is NULL.}

\item{metrics}{A character vector specifying the desired metrics to compute. Options include:
- `"R_square"`: \( R^2 \) metric.
- `"L_square"`: \( L^2 \) metric.
- `"Pesudo_R"`: Pseudo-\( R^2 \) metric.
- `"Harrell’s C"`: Harrell's Concordance Index.
- `"Uno’s C"`: Uno's Concordance Index.
- `"R_sph"`: Explained variation (\( R_{sph} \)).
- `"R_sh"`: Explained variation (\( R_{sh} \)).
- `"Brier Score"`: Brier Score.
- `"Time Dependent Auc"`: Time-dependent AUC.}
}
\value{
A data frame summarizing the requested performance metrics.
}
\description{
This function computes and summarizes various performance metrics for survival data
using predicted values and observed survival data. Users can specify the desired metrics
or compute all available metrics by default.
}
\examples{
predicted_data <- c(0.8, 0.6, 0.4, 0.2)
survival_time <- c(5, 8, 3, 10)
status <- c(1, 0, 1, 1)
metric <- "R_square"
pam.metrics_summary_predicted(predicted_data, survival_time, metric, status)

}
