% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.re.R
\name{pam.re}
\alias{pam.re}
\title{RE Measure of Explained Variation of the Model}
\usage{
pam.re(fit, ...)
}
\arguments{
\item{fit}{A survival model object (e.g., \code{coxph}, \code{survreg}, \code{aareg}).}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list containing the various RE measures and their associated statistics.
}
\description{
This package provides tools to evaluate the proportion of variability in survival times explained by covariates using Schempers' RE measure. The measure assesses how well the model explains the variability in the data, which is important for understanding the impact of covariates in survival analysis.

The RE measure is sourced from:
\url{https://ibmi3.mf.uni-lj.si/ibmi-english/biostat-center/programje/Re.r}

It includes the following functions:

\itemize{
  \item \code{re} - Function for calculating the RE measure.
  \item \code{summary.re} - Function for calculating the RE measure over time (cumulative and in specified time windows).
}

The \code{re} function can be called with:

\code{re(fit)}

It works with various model objects, including \code{coxph}, \code{survreg}, and \code{aareg}.
}
\details{
The \code{re} function calculates the RE measure, which consists of the following components:
- \strong{Re}: The primary RE measure.
- \strong{Re.imp}: The RE measure assuming constant coefficients after the last event time.
- \strong{Re.fix}: The RE measure assuming a constant value of coefficients throughout the follow-up time. (Note: Returns NA for counting process type times.)
- \strong{se}: Standard error of the RE measure.
- \strong{sen0}: Standard error calculated with expected Q.
- \strong{C}: C-index generalized to allow for time-dependent covariates, ensuring independence from the censoring process.
- \strong{r2nw}: An unweighted measure.
}
\note{
Left-truncated data is not yet implemented. When calculating \code{Re.imp}, the dataset should not be split after the last event time; each data line should represent a different subject.
}
\examples{
# Load necessary libraries
library(survival)
library(PAmeasures)
# Use Mayo Clinic Primary Biliary Cirrhosis Data
data(pbc)

pbc <- pbc \%>\% 
  filter(!is.na(trt)) \%>\% 
  mutate(log_albumin = log(albumin),
         log_bili = log(bili),
         log_protime = log(protime),
         status = ifelse(status == 2, 1, 0))

# Fit a full Cox PH model
fit.coxph.full <- coxph(Surv(time, status) ~ age + log_albumin + 
                         log_bili + log_protime + edema, 
                         data = pbc, x = TRUE, y = TRUE)
# Calculate the RE measure
re_result <- pam.re(fit.coxph.full)
print(re_result$Re)

# Summarize the RE measure over time
summary_result <- pam.summary(re_result)

}
\references{
- Schemper, M. and R. Henderson (2000). Predictive accuracy and explained variation in Cox regression. Biometrics 56, 249--255.
- LUSA, L., MICELI, R. and MARIANI, L. (2007). Estimation of predictive accuracy in survival analysis using R and S-PLUS. Computer methods and programs in biomedicine 87 132–137.
}
