% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.rsh_metric.R
\name{pam.rsh_metric}
\alias{pam.rsh_metric}
\title{Compute R_sh Metric}
\usage{
pam.rsh_metric(predicted_data, survival_time, status)
}
\arguments{
\item{predicted_data}{A numeric vector of predicted survival probabilities.}

\item{survival_time}{A numeric vector of observed survival times.}

\item{status}{A numeric vector indicating event occurrence (1 for event, 0 for censoring).}
}
\value{
A list with the following components:
\item{D}{The total variation in the survival data.}
\item{Dx}{The unexplained variation by the predictions.}
\item{V}{The explained variation (\( R_{sh} \)).}
}
\description{
This function calculates the \( R_{sh} \) metric, a measure of explained variation for survival models,
using only predicted survival probabilities and observed survival data.
}
\examples{
library(PAmeasure)
predicted_data <- c(0.8, 0.6, 0.4, 0.2)
survival_time <- c(5, 8, 3, 10)
status <- c(1, 0, 1, 1)
pam.rsh_metric(predicted_data, survival_time, status)

}
\references{
Schemper, M. and R. Henderson (2000). Predictive accuracy and explained variation in Cox regression.
Biometrics 56, 249--255.

Lusa, L., R. Miceli and L. Mariani (2007). Estimation of predictive accuracy in survival analysis
using R and S-PLUS. Computer Methods and Programs in Biomedicine 87, 132--137.

Potapov, S., Adler, W., Schmid, M., Bertrand, F. (2024). survAUC: Estimating Time-Dependent AUC for Censored Survival Data. 
R package version 1.3-0. DOI: \doi{10.32614/CRAN.package.survAUC}. Available at \url{https://CRAN.R-project.org/package=survAUC}.
}
