% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.schemper.R
\name{pam.schemper}
\alias{pam.schemper}
\title{Distance-based estimator of survival predictive accuracy proposed by Schemper and Henderson}
\usage{
pam.schemper(train.fit, traindata, newdata)
}
\arguments{
\item{train.fit}{A fitted Cox model from `rms::cph`.}

\item{traindata}{A data frame containing training data.}

\item{newdata}{A data frame with new data for prediction.}
}
\value{
A list containing the model call, D, Dx, V metrics, and other calculated components.
}
\description{
This function calculates metrics (D, Dx, V) based on a fitted Cox model using the `cph` function from
the `rms` package. It uses survival estimates to assess the concordance between observed and predicted
survival outcomes.
}
\examples{
library(PAmeasures)
library(survival)

# Use Mayo Clinic Primary Biliary Cirrhosis Data
data(pbc)
pbc <- pbc \%>\% 
  filter(is.na(trt)==F) \%>\% 
mutate(log_albumin = log(albumin),
       log_bili = log(bili),
       log_protime = log(protime),
       status = ifelse(status==2, 1, 0))
#Schemper and Henderson's estimator of the absolute deviation between survival functions
schemper(train.fit.full, pbc, pbc)$Dx       
}
\references{
Schemper, M. and R. Henderson (2000). Predictive accuracy and explained variation in Cox regression.
Biometrics 56, 249--255.

Lusa, L., R. Miceli and L. Mariani (2007). Estimation of predictive accuracy in survival analysis
using R and S-PLUS. Computer Methods and Programs in Biomedicine 87, 132--137.

Potapov, S., Adler, W., Schmid, M., Bertrand, F. (2024). survAUC: Estimating Time-Dependent AUC for Censored Survival Data. 
R package version 1.3-0. DOI: \doi{10.32614/CRAN.package.survAUC}. Available at \url{https://CRAN.R-project.org/package=survAUC}.
}
