% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.survivalROC.R
\name{pam.survivalROC}
\alias{pam.survivalROC}
\title{Time-dependent ROC Curve from Censored Survival Data}
\usage{
pam.survivalROC(
  Stime,
  status,
  marker,
  entry = NULL,
  predict.time,
  cut.values = NULL,
  method = "NNE",
  lambda = NULL,
  span = NULL,
  window = "symmetric"
)
}
\arguments{
\item{Stime}{Event time or censoring time for subjects.}

\item{status}{Indicator of status, 1 if death or event, 0 otherwise.}

\item{marker}{Predictor or marker value.}

\item{entry}{Entry time for the subjects.}

\item{predict.time}{Time point of the ROC curve.}

\item{cut.values}{Marker values to use as a cut-off for calculating sensitivity and specificity.}

\item{method}{Method for fitting joint distribution of (marker, t), either "KM" or "NNE". Default is "NNE".}

\item{lambda}{Smoothing parameter for NNE.}

\item{span}{Span for the NNE. Either \code{lambda} or \code{span} is required for NNE.}

\item{window}{Window type for NNE, either "symmetric" or "asymmetric".}
}
\value{
A list containing the following elements:
\describe{
  \item{cut.values}{Unique marker values for calculating TP and FP.}
  \item{TP}{True Positive corresponding to the cut-offs in the marker.}
  \item{FP}{False Positive corresponding to the cut-offs in the marker.}
  \item{predict.time}{Time point of interest.}
  \item{Survival}{Kaplan-Meier survival estimate at \code{predict.time}.}
  \item{AUC}{Area Under the ROC Curve at \code{predict.time}.}
}
}
\description{
This function creates a time-dependent ROC curve from censored survival data
using the Kaplan-Meier (KM) or Nearest Neighbor Estimation (NNE) method of
Heagerty, Lumley, and Pepe (2000).
}
\examples{
library(survival)
library(PAmeasures)
# Use Mayo Clinic Primary Biliary Cirrhosis Data
data(pbc)
pbc <- pbc \%>\% 
  filter(is.na(trt)==F) \%>\% 
  mutate(log_albumin = log(albumin),
         log_bili = log(bili),
         log_protime = log(protime),
         status = ifelse(status==2, 1, 0))

time_dep_auc_full <- survivalROC(Stime = pbc$time,
            status = pbc$status,
            marker = predict(fit.coxph.full, newdata=pbc, type = "lp") ,
            predict.time = quantile(pbc$time, 0.5), 
            method="KM")$AUC
}
\references{
Heagerty, P. J., Lumley, T., & Pepe, M. S. (2000). Time-dependent ROC curves for censored
survival data and a diagnostic marker. \emph{Biometrics}, 56(2), 337-344.

Heagerty, P. J., Saha-Chaudhuri, P. (2022). survivalROC: Time-Dependent ROC Curve Estimation from Censored Survival Data. 
R package version 1.0.3.1. DOI: \doi{10.32614/CRAN.package.survivalROC}. Available at \url{https://CRAN.R-project.org/package=survivalROC}.
}
