% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.survreg.metrics.R
\name{pam.survreg.metrics}
\alias{pam.survreg.metrics}
\title{Prediction Accuracy Measures for Predicted Survival Data}
\usage{
pam.survreg.metrics(observed_data, predicted_data, dist)
}
\arguments{
\item{observed_data}{A data frame containing the observed survival data. Must include columns:
- \code{time}: Observed survival times.
- \code{status}: Censoring indicator (1 = event occurred, 0 = censored).}

\item{predicted_data}{A numeric vector of predicted survival times or risk scores. Must have the same length as \code{observed_data}.}
}
\value{
A list containing:
  - \code{R.squared}: The \(R^2\) metric.
  - \code{L.squared}: The \(L^2\) metric.
  - \code{Pseudo.R}: The pseudo \(R^2\) metric, computed as the product of \(R^2\) and \(L^2\).
}
\description{
Computes \(R^2\), \(L^2\), and Pseudo \(R^2\) metrics for predicted survival data.. R-squared is an extension of the classical R2 statistic for a linear model, quantifying the amount of variability in the response that is explained by a corrected prediction based on the original prediction function. L-squared is the proportion of the prediction error of the original prediction function that is explained by the corrected prediction function, quantifying the distance between the corrected and uncorrected predictions. When used together, they give a complete summary of the predictive power of a prediction function.
}
\examples{
library(survival)
library(PAmeasures)

# Use Mayo Clinic Primary Biliary Cirrhosis Data
data(pbc)

head(pbc)

}
