% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.coxph_restricted.R
\name{pam.coxph_restricted}
\alias{pam.coxph_restricted}
\title{Prediction Accuracy Measures for Cox Proportional Hazards Model}
\usage{
pam.coxph_restricted(model, covs, tau = NULL, new_data = NULL, predict = T)
}
\arguments{
\item{model}{An object of class `coxph` representing a fitted Cox proportional hazards regression model. The model must be fitted with `x = TRUE` and `y = TRUE` to include the design matrix and response vector.}

\item{covs}{A character vector specifying the names of the covariates used in the model.}

\item{tau}{(Optional) A numeric value specifying a time horizon for truncating the observed survival times. If provided, the function calculates metrics up to time `tau`.}

\item{new_data}{(Optional) A new dataset for validation. If provided, the function calculates metrics on this new dataset instead of the training data.}

\item{predict}{If true, return a list contains predicted survival probabilities.}
}
\value{
A list containing the following components:
\item{R.squared}{The R-squared measure, quantifying the proportion of variability explained by the model.}
\item{L.squared}{The L-squared measure, quantifying the proportion of prediction error explained by the corrected prediction.}
\item{Psuedo.R}{A pseudo-R measure, calculated as the product of R-squared and L-squared.}
}
\description{
This function calculates a set of measures, $R^2$ and $L^2$, $Pseudo-R^2$, for a Cox proportional hazards model. When predict = TRUE, 
it returns the predicted survival times alongside observed values and event status. 
R-squared quantifies the proportion of variability in the observed survival times explained by the model's predictions,
while L-squared measures the proportion of prediction error explained by a corrected prediction function. Together, these metrics provide a comprehensive evaluation of the predictive power of the Cox model.
}
\examples{
library(survival)
library(TimeMetric)
library(tidyverse)
set.seed(2025)
data(pbc, package = "survival")
df <- pbc \%>\% 
 filter(is.na(trt)==F) \%>\% 
 mutate(log_albumin = log(albumin),
        log_bili = log(bili),
        log_protime = log(protime),
        status = ifelse(status==2, 1, 0)) \%>\% 
 select(time, status, age, log_albumin, log_bili, log_protime, edema)
train_data_idx <- sample(1:dim(df)[1], round(2/3*dim(df)[1]))
train_data <- df[train_data_idx, ]
test_data <- df[-train_data_idx, ]
m.cox<- coxph(Surv(time, status) ~ .,
             data = train_data, x = TRUE, y = TRUE)
m.wei <- survreg(Surv(time, status) ~ ., #scale = 0.1,
                data = train_data, dist="weibull", x=TRUE, y=TRUE)
m.wei.fix <- survreg(Surv(time, status) ~ .,  scale = 5,
                    data = train_data, dist="weibull", x=TRUE, y=TRUE)
## predict prob
covs <- names(df)[-c(1:2)]
cox_pred <- pam.coxph_restricted(model = m.cox,
                                covs = covs,
                                new_data = test_data) 
}
\references{
Li, G., & Wang, X. (2016). Prediction Accuracy Measures for a Nonlinear Model and for Right-Censored Time-to-Event Data.
arXiv preprint arXiv:1611.03063. Available at \url{https://arxiv.org/abs/1611.03063}.
}
