% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.predicted_survial_eval_two_phase.R
\name{pam.predicted_survial_eval_two_phase}
\alias{pam.predicted_survial_eval_two_phase}
\alias{pam.predicted_survial_eval_casecohort}
\alias{pam.predicted_survial_eval_ncc}
\title{Evaluate Survival Model Metrics for Two-Phase Sampling Designs}
\usage{
pam.predicted_survial_eval_two_phase(
  analysis_data,
  pred_results,
  eval_time,
  km_cens_fit,
  case_weights,
  metrics = c("Harrell’s C", "Uno’s C", "Brier Score", "Time Dependent Auc",
    "R_square", "L_square", "Pesudo_R")
)

pam.predicted_survial_eval_casecohort(
  analysis_data,
  pred_results,
  eval_time,
  km_cens_fit,
  case_weights,
  metrics = c("Harrell’s C", "Uno’s C", "Brier Score", "Time Dependent Auc",
    "R_square", "L_square", "Pesudo_R")
)

pam.predicted_survial_eval_ncc(
  analysis_data,
  pred_results,
  eval_time,
  km_cens_fit,
  case_weights,
  metrics = c("Harrell’s C", "Uno’s C", "Brier Score", "Time Dependent Auc",
    "R_square", "L_square", "Pesudo_R")
)
}
\arguments{
\item{analysis_data}{A data frame containing the analysis dataset used in the fitted Cox model.
Must include `time` and `status` columns.}

\item{pred_results}{A numeric vector of predicted survival probabilities with (can be calculate by `pam.coxph_restricted()`) :
- `Prob`: predicted survival probabilities over time.
- `pred`: predicted risk scores.
- `time`: time points for `Prob` rows.}

\item{eval_time}{Time horizon at which to evaluate performance.}

\item{case_weights}{Numeric vector of case weights (from `weighted_param()`).}

\item{km_cens}{Kaplan–Meier estimate of censoring distribution (`survival::survfit` object).}
}
\value{
A tibble with columns:
  - `Metric`: name of metric
  - `Value`: numeric value
}
\description{
This function constructs a standardized evaluation dataset (`dat1`) from the
analysis dataset, predicted survival probabilities, and sampling weights, and
computes key performance metrics in a tidy table format.
Works for common two-phase sampling designs:
  - Case–cohort (unstratified or stratified)
  - Nested case–control (matched or unmatched)
}
\keyword{internal}
