% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_survival_eval_cr.R
\name{tm_survival_eval_cr}
\alias{tm_survival_eval_cr}
\title{Evaluate Survival Prediction Performance with Competing Risks}
\usage{
tm_survival_eval_cr(
  pred_cif,
  event_time,
  time.cif,
  status,
  metrics = NULL,
  t_star = NULL,
  tau = NULL,
  event_type = 1
)
}
\arguments{
\item{pred_cif}{A matrix of predicted cumulative incidence functions (CIFs) for different subjects.}

\item{event_time}{A numeric vector of observed event or censoring times for subjects.}

\item{time.cif}{A numeric vector of time points corresponding to the predicted CIFs.}

\item{status}{A numeric vector indicating the event type for each subject: by default
\itemize{
  \item `0` for censored cases.
  \item `1` for the primary event of interest.
  \item `2` (or other values) for competing events.
}}

\item{metrics}{A character vector specifying the performance metrics to calculate. 
The default is `"all"`, which computes all available metrics. 
Valid options include:
\itemize{
  \item `"pseudo_r2"`
  \item `"r_square"`
  \item `"l_square"`
  \item `"pseudo_r2_point"`
  \item `"r2_point"`
  \item `"l2_point"`
  \item `"c_index"`
  \item `"brier_score"`
  \item `"td_auc"`
}}

\item{t_star}{Optional numeric value specifying a reference time for evaluating point-version Pseudo_R_square, 
the Brier Score and Time-Dependent AUC. Defaults to the median of `event_time`.}

\item{tau}{Optional numeric value specifying a restricted time horizon for model evaluation. 
Defaults to the maximum observed event time plus one.}

\item{event_type}{An integer indicating the primary event type of interest (default = `1`).}
}
\value{
A data frame with two columns:
  \itemize{
    \item `"Metric"`: Name of the evaluated metric.
    \item `"Value"`: Computed value of the corresponding metric.
  }
}
\description{
This function evaluates the predictive performance of a survival model in the presence of competing risks.
It calculates various performance metrics including Pseudo R-squared, R-squared, L-squared, C-index, 
Brier Score, and Time-Dependent AUC.
}
